requirements.txt
Django==3.2.24
djangorestframework==3.14.0
boto3==1.23.10
django-storages==1.12.3
mysqlclient==2.1.1
python-dateutil==2.8.2